moved tests into designated folder
